requirements.txt
brotli==1.0.9
python-lorem==1.1.2